Serialise multiple TPM accesses to the same file descriptor
The Linux tpm interface allows for concurrent access if each writer opens
the device node separately. However, if multiple threads use a shared
reference to the device node, commands will fail if submitted while another
command is in-flight. The simplest approach to deal with this is simply to
add a mutex in this scenario and use that during RunCommand().